Added new closed set data generator
