+++ auth_ident_requirements.txt
@@ -15,3 +15,4 @@ numpydoc >= 0.7.0
 pandas >= 0.24.2
 bs4
 tables
+sklearn
